worker/fileconversor/requirements.txt
celery[redis]
psycopg2-binary
requests
moviepy
sqlalchemy==1.4.40
python-dotenv
google-cloud-pubsub
google-cloud-storage
Flask